Store temperature sensor value as a float.

+++ temperature_sensor.h
@@ -2,7 +2,7 @@
 #define TEMPERATURE_SENSOR_H
 
 namespace TemperatureSensor {
-  extern int temperature;
+  extern float temperature;
 
   void setup();
   void update();
